rename comps
