+++ apps/domain-clients-example/src/components/cosmos/CosmosDenomTransfer.tsx
@@ -2,58 +2,46 @@
 
 import { Button, Input, Label } from '@/components/ui';
 import { useCallback, useState } from 'react';
-import { parseUnits } from 'viem';
-import {
-  useSigningEvmClient,
-  useEvmClient,
-} from '@valence-protocol/domain-clients-react/evm';
+import { useSigningCosmosClient } from '@valence-protocol/domain-clients-react/cosmos';
 import { useMutation } from '@tanstack/react-query';
-import { useSwitchChain } from 'wagmi';
-import { useIsEvmChainConnected } from '@valence-protocol/domain-modal-react';
+import { baseToMicro } from '@valence-protocol/domain-clients-core';
+import { useIsCosmosChainConnected } from '@valence-protocol/domain-modal-react';
 
-type EthereumTestnetWriteProps = {
-  chainId: number;
+type CosmosDenomTransferProps = {
+  chainId: string;
+  decimals: number;
+  denom: string;
 };
 
-const sepoliaEthDecimals = 18;
-
-export const EthereumTestnetWrite = ({
+export const CosmosDenomTransfer = ({
   chainId,
-}: EthereumTestnetWriteProps) => {
+  decimals,
+  denom,
+}: CosmosDenomTransferProps) => {
   const [toAddress, setToAddress] = useState('');
   const [amount, setAmount] = useState('');
-  const { switchChain } = useSwitchChain();
-  const isChainConnected = useIsEvmChainConnected(chainId);
+  const isConnected = useIsCosmosChainConnected({ chainId });
 
-  const signingEvmClient = useSigningEvmClient(chainId);
-  const publicEvmClient = useEvmClient(chainId);
+  const signingCosmosClient = useSigningCosmosClient({ chainId });
 
   const onSubmit = useCallback(async () => {
-    switchChain({ chainId: chainId });
-    if (!signingEvmClient || !publicEvmClient) {
-      throw new Error('EVM client not found');
+    if (!signingCosmosClient) {
+      throw new Error('Cosmos signing client not found');
     }
-    const amountInWei = parseUnits(amount, sepoliaEthDecimals);
 
-    const tx = await signingEvmClient.sendTokens({
-      to: toAddress as `0x${string}`,
-      value: amountInWei,
+    const amountInBase = baseToMicro(amount, decimals);
+
+    const tx = await signingCosmosClient.sendTokens({
+      recipient: toAddress,
+      amount: [
+        {
+          amount: amountInBase.toString(),
+          denom,
+        },
+      ],
     });
-    const txHash = await publicEvmClient.waitForTransactionReceipt({
-      txHash: tx,
-    });
-    if (txHash.status === 'success') {
-      return txHash;
-    }
-    throw new Error('Transaction failed');
-  }, [
-    amount,
-    toAddress,
-    signingEvmClient,
-    publicEvmClient,
-    chainId,
-    switchChain,
-  ]);
+    return tx;
+  }, [decimals, denom, amount, toAddress, signingCosmosClient]);
 
   const {
     mutate: sendTokens,
@@ -68,9 +56,9 @@ export const EthereumTestnetWrite = ({
   });
 
   return (
-    <div className='flex flex-col gap-2'>
+    <div className='flex flex-col gap-2 '>
       <div className='flex flex-col'>
-        <Label htmlFor='amount'>Amount SepoliaETH</Label>
+        <Label htmlFor='amount'>Amount NTRN</Label>
         <Input
           className='border border-gray-300 rounded-sm p-1 font-mono text-xs max-w-[400px]'
           placeholder='Enter amount'
@@ -90,13 +78,13 @@ export const EthereumTestnetWrite = ({
         />
       </div>
       <div className='flex flex-row gap-2 items-center'>
-        <Button disabled={!isChainConnected} onClick={() => sendTokens()}>
+        <Button disabled={!isConnected} onClick={() => sendTokens()}>
           <span>Transfer</span>
         </Button>
-        {!isChainConnected && (
-          <p className='text-xs text-gray-500'>
-            Connect to Sepolia to transfer tokens
-          </p>
+        {!isConnected && (
+          <div className='text-xs text-gray-500'>
+            Connect to Neutron Testnet to transfer NTRN
+          </div>
         )}
       </div>
       {isError && (
