+++ apps/domain-clients-example/src/components/cosmos/index.ts
@@ -1,4 +1,4 @@
-export * from './NeutronRead';
-export * from './NeutronTestnetWrite';
+export * from './CosmosDenomBalance';
+export * from './CosmosDenomTransfer';
 export * from './ConnectCosmosChain';
 export * from './CosmosOps';
